- Looping playback. Has some known glitches but nothing fatal. - Some old changes, dunno what

+++ Sound.h
@@ -11,8 +11,8 @@
 @interface Sound : NSManagedObject {
 	bool filter_on;// = false
 }
-+(NSInteger)countInContext:(NSManagedObjectContext*)ctx;
-+(NSInteger)highestIndexInContext:(NSManagedObjectContext*)ctx;
++(int)countInContext:(NSManagedObjectContext*)ctx;
++(int)highestIndexInContext:(NSManagedObjectContext*)ctx;
 
 -(void)generateParamsFromCategory:(NSString*)templateName;
