+++ common.h
@@ -1,6 +1,6 @@
 #define rnd(n) (rand()%(n+1))
 extern float frnd(float range);
 
-inline NSError *NSMakeError(NSString *domain, NSInteger code, NSString *descr, NSString *recoveryDescr) { return [NSError errorWithDomain:domain code:code userInfo:[NSDictionary dictionaryWithObjectsAndKeys:descr, NSLocalizedDescriptionKey, recoveryDescr, NSLocalizedRecoverySuggestionErrorKey, nil]]; }
+inline NSError *NSMakeError(NSString *domain, int code, NSString *descr, NSString *recoveryDescr) { return [NSError errorWithDomain:domain code:code userInfo:[NSDictionary dictionaryWithObjectsAndKeys:descr, NSLocalizedDescriptionKey, recoveryDescr, NSLocalizedRecoverySuggestionErrorKey, nil]]; }
 
-inline NSError *CfxrMakeError(NSInteger code, NSString *descr, NSString *recoveryDescr) { return NSMakeError(@"eu.thirdcog.cfxr.ErrorDomain", code, descr, recoveryDescr); }
+inline NSError *CfxrMakeError(int code, NSString *descr, NSString *recoveryDescr) { return NSMakeError(@"eu.thirdcog.cfxr.ErrorDomain", code, descr, recoveryDescr); }
